feat: Add RPC command session helpers and methods
