+++ packages/prime-sandboxes/src/prime_sandboxes/sandbox.py
@@ -9,15 +9,13 @@ import sys
 import time
 import uuid
 from datetime import datetime, timezone
-from typing import Any, Dict, List, NoReturn, Optional, Protocol, cast
+from typing import Any, Dict, List, NoReturn, Optional
 
 import aiofiles
 import httpx
 from connectrpc.client import ConnectClient, ConnectClientSync
 from connectrpc.code import Code
 from connectrpc.errors import ConnectError
-from connectrpc.method import IdempotencyLevel, MethodInfo
-from google.protobuf.message import Message
 from tenacity import (
     retry,
     retry_if_exception_type,
@@ -25,7 +23,6 @@ from tenacity import (
     wait_exponential,
 )
 
-from ._proto.command_session import command_session_pb2
 from .core import APIClient, APIError, AsyncAPIClient
 from .exceptions import (
     CommandTimeoutError,
@@ -54,6 +51,11 @@ from .models import (
     SandboxLogsResponse,
     SSHSession,
 )
+from .rpc_command_session import (
+    COMMAND_SESSION_START_RPC_METHOD,
+    build_command_session_start_request,
+    collect_command_session_start_event,
+)
 
 # Retry configuration for transient connection errors on gateway requests
 # Note: ReadTimeout is NOT included because the request may have been processed
@@ -77,64 +79,6 @@ _gateway_retry = retry(
 
 
 _ENV_VAR_PATTERN = re.compile(r"^[A-Za-z_][A-Za-z0-9_]*$")
-
-
-class _CommandSpecLike(Protocol):
-    cwd: str
-
-
-class _CommandSpecFactory(Protocol):
-    def __call__(self, *, cmd: str, args: List[str], envs: Dict[str, str]) -> _CommandSpecLike: ...
-
-
-class _CommandSessionStartRequestFactory(Protocol):
-    def __call__(self, *, command: _CommandSpecLike, stdin: bool) -> Message: ...
-
-
-class _CommandSessionDataEventLike(Protocol):
-    stdout: bytes
-    stderr: bytes
-    pty: bytes
-
-    def WhichOneof(self, field_name: str) -> str | None: ...
-
-
-class _CommandSessionEndEventLike(Protocol):
-    exit_code: int
-
-
-class _CommandSessionEventLike(Protocol):
-    data: _CommandSessionDataEventLike
-    end: _CommandSessionEndEventLike
-
-    def WhichOneof(self, field_name: str) -> str | None: ...
-
-
-class _CommandSessionStartResponseLike(Protocol):
-    event: _CommandSessionEventLike
-
-    def HasField(self, field_name: str) -> bool: ...
-
-
-_COMMAND_SESSION_START_REQUEST_TYPE = cast(
-    type[Message], getattr(command_session_pb2, "StartRequest")
-)
-_COMMAND_SESSION_START_RESPONSE_TYPE = cast(
-    type[Message], getattr(command_session_pb2, "StartResponse")
-)
-_COMMAND_SESSION_START_REQUEST_FACTORY = cast(
-    _CommandSessionStartRequestFactory, _COMMAND_SESSION_START_REQUEST_TYPE
-)
-_COMMAND_SPEC_FACTORY = cast(_CommandSpecFactory, getattr(command_session_pb2, "CommandSpec"))
-
-
-_COMMAND_SESSION_START_RPC_METHOD = MethodInfo(
-    name="Start",
-    service_name="command_session.CommandSession",
-    input=_COMMAND_SESSION_START_REQUEST_TYPE,
-    output=_COMMAND_SESSION_START_RESPONSE_TYPE,
-    idempotency_level=IdempotencyLevel.UNKNOWN,
-)
 
 
 def _build_user_agent() -> str:
@@ -191,47 +135,6 @@ def _is_gateway_sandbox_not_found(response: Optional[httpx.Response]) -> bool:
         return False
 
     return body.get("error") == "sandbox_not_found"
-
-
-def _build_command_session_start_request(
-    command: str,
-    working_dir: Optional[str],
-    env: Optional[Dict[str, str]],
-) -> Message:
-    command_spec = _COMMAND_SPEC_FACTORY(
-        cmd="/bin/bash",
-        args=["-l", "-c", command],
-        envs=env or {},
-    )
-    if working_dir is not None:
-        command_spec.cwd = working_dir
-
-    return _COMMAND_SESSION_START_REQUEST_FACTORY(command=command_spec, stdin=False)
-
-
-def _collect_command_session_start_event(
-    response: _CommandSessionStartResponseLike,
-    stdout_parts: List[str],
-    stderr_parts: List[str],
-) -> Optional[int]:
-    if not response.HasField("event"):
-        return None
-
-    event = response.event
-    event_kind = event.WhichOneof("event")
-
-    if event_kind == "data":
-        data_kind = event.data.WhichOneof("output")
-        if data_kind == "stdout" and event.data.stdout:
-            stdout_parts.append(event.data.stdout.decode("utf-8", errors="replace"))
-        elif data_kind == "stderr" and event.data.stderr:
-            stderr_parts.append(event.data.stderr.decode("utf-8", errors="replace"))
-        elif data_kind == "pty" and event.data.pty:
-            stdout_parts.append(event.data.pty.decode("utf-8", errors="replace"))
-    elif event_kind == "end":
-        return int(event.end.exit_code)
-
-    return None
 
 
 def _raise_not_running_error(
@@ -638,7 +541,7 @@ class SandboxClient:
         base_url = f"{gateway_url}/{auth['user_ns']}/{auth['job_id']}"
         headers = {"Authorization": f"Bearer {auth['token']}"}
         effective_timeout = timeout if timeout is not None else 300
-        request = _build_command_session_start_request(command, working_dir, env)
+        request = build_command_session_start_request(command, working_dir, env)
         stdout_parts: List[str] = []
         stderr_parts: List[str] = []
         exit_code: Optional[int] = None
@@ -647,13 +550,13 @@ class SandboxClient:
         try:
             stream = rpc_client.execute_server_stream(
                 request=request,
-                method=_COMMAND_SESSION_START_RPC_METHOD,
+                method=COMMAND_SESSION_START_RPC_METHOD,
                 headers=headers,
                 timeout_ms=effective_timeout * 1000,
             )
             for event in stream:
-                event_exit_code = _collect_command_session_start_event(
-                    cast(_CommandSessionStartResponseLike, event),
+                event_exit_code = collect_command_session_start_event(
+                    event,
                     stdout_parts,
                     stderr_parts,
                 )
@@ -1393,7 +1296,7 @@ class AsyncSandboxClient:
         base_url = f"{gateway_url}/{auth['user_ns']}/{auth['job_id']}"
         headers = {"Authorization": f"Bearer {auth['token']}"}
         effective_timeout = timeout if timeout is not None else 300
-        request = _build_command_session_start_request(command, working_dir, env)
+        request = build_command_session_start_request(command, working_dir, env)
         stdout_parts: List[str] = []
         stderr_parts: List[str] = []
         exit_code: Optional[int] = None
@@ -1402,13 +1305,13 @@ class AsyncSandboxClient:
         try:
             stream = rpc_client.execute_server_stream(
                 request=request,
-                method=_COMMAND_SESSION_START_RPC_METHOD,
+                method=COMMAND_SESSION_START_RPC_METHOD,
                 headers=headers,
                 timeout_ms=effective_timeout * 1000,
             )
             async for event in stream:
-                event_exit_code = _collect_command_session_start_event(
-                    cast(_CommandSessionStartResponseLike, event),
+                event_exit_code = collect_command_session_start_event(
+                    event,
                     stdout_parts,
                     stderr_parts,
                 )
